Remove map, reduce, walk - use Reducers instead
Reducers are modules which implement some traversal of a graph,
they are intended to work much in the same vein as `Enum.reduce/3`.
The old API for mapping/reducing/walking was not ideal for a number
of reasons, the Reducer API is more extensible.

There are two reducers out of the box, `Graph.Reducers.Bfs` and
`Graph.Reducers.Dfs` for breadth-first and depth-first traversals
respectively.

I would still like a way to unify traversals, but have not come up with
an API that feels right just yet.